virgl: add blob flag to create new guest blob resource
Add a new resource flag for guest blob: VIRGL_RENDERER_BLOB_FLAG_USE_USERPTR.

Different with the old guest resource, it is mapped from guest,
and it is a blob memory that need use get_blob to bypass the resource
to low level context in native context. This is for feature like
OpenCL userptr.

Change-Id: I73cfcf511363d5382def07ea4c80dfd2f2a824cf
Signed-off-by: Honglei Huang <Honglei1.Huang@amd.com>